requirements.txt
streamlit
numpy 
pickle
PyMuPDF
python-docx
google.generativeai
faiss-cpu
scikit-learn